+++ VotingApp/VotingApp.qbs
@@ -12,10 +12,6 @@ QtGuiApplication {
     cpp.includePaths: [".", "qml-promise/src"]
 
     files: [
-        "ChainAdaptorWrapper.cpp",
-        "ChainAdaptorWrapper.hpp",
-        "BackendWrapper.cpp",
-        "BackendWrapper.hpp",
         "PromiseConverter.cpp",
         "PromiseConverter.hpp",
         "TwoPartyClient.cpp",
@@ -29,6 +25,10 @@ QtGuiApplication {
         "main.cpp",
         "qml/*.qml",
         "qml/qml.qrc",
+        "wrappers/ChainAdaptorWrapper.cpp",
+        "wrappers/ChainAdaptorWrapper.hpp",
+        "wrappers/BackendWrapper.cpp",
+        "wrappers/BackendWrapper.hpp",
         "wrappers/Balance.cpp",
         "wrappers/Balance.hpp",
         "wrappers/Coin.cpp",
