+++ VotingApp/VotingSystem.cpp
@@ -27,11 +27,11 @@
 #include "wrappers/Contest.hpp"
 #include "wrappers/Decision.hpp"
 #include "wrappers/Datagram.hpp"
-#include "BackendWrapper.hpp"
+#include "wrappers/BackendWrapper.hpp"
 #include "wrappers/OwningWrapper.hpp"
 #include "wrappers/Converters.hpp"
+#include "wrappers/ChainAdaptorWrapper.hpp"
 #include "VotingSystem.hpp"
-#include "ChainAdaptorWrapper.hpp"
 #include "Promise.hpp"
 #include "PromiseConverter.hpp"
 #include "TwoPartyClient.hpp"
