+++ VotingApp/wrappers/BackendWrapper.cpp
