+++ VotingApp/wrappers/ChainAdaptorWrapper.cpp
